Writing Tools/broken_ones/requirements_pinecone.txt
openai>=1.0.0
pinecone-client[grpc]>=3.0.0
tiktoken>=0.5.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
langchain>=0.0.300
langchain-text-splitters>=0.0.1
docx2txt>=0.8
numpy>=1.20.0 